requirements.txt
asgiref==3.7.2
Django==4.2.4
psycopg2==2.9.7
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
Pillow==10.0.0
django-phonenumber-field==7.1.0
phonenumbers-8.13.21
